Personalizzata la GUI per finestra temporale e range di valori dei campioni Modificato il vertex shader per il la conversione tra campioni e posizione vertici nel vertex stage La lettura dei campioni nell'applicazione ora è solo di uint16_t

+++ include/utils.h
@@ -1,6 +1,6 @@
 #include <mlcontext.h>
 
-typedef float sampleType;
+typedef uint16_t sampleType;
 
 //Struttura dati per il passaggio di argomenti
 struct argsThreadStruct{
